wam, chromium53: point to eel branches.
To avoid misleading users of the layers for the WAM port for AGL,
let's specify the AGL version each branch is for. For chromium53
and wam, we move from using @1.agl to @1.agl.eel, to follow
same convention we are using for flounder.

We should eventually drop master branch in meta-agl-lge and set
the default branch to be the one actively being developed.

[SPEC-1867] Rename webosose eel branches from master to eel