mesa: add patches to fix swrast resize.
Launching web apps on qemux86 are crashing in the GPU thread. After
debugging, it seems the problem seems to be related to not properly
handling surface resize (and first size) for rendering.

Backported the 3 patches solving the problem upstream:

    From: "Juan A. Suarez Romero" <jasuarez@igalia.com>
    Date: Mon, 4 Jun 2018 10:22:49 +0000
    Subject: wayland/egl: initialize window surface size to window size

    From: "Juan A. Suarez Romero" <jasuarez@igalia.com>
    Date: Wed, 6 Jun 2018 10:13:05 +0000
    Subject: wayland/egl: update surface size on window resize

    From: Olivier Fourdan <ofourdan@redhat.com>
    Date: Thu, 25 Oct 2018 14:48:15 +0200
    Subject: wayland/egl: Resize EGL surface on update buffer for swrast

These avoid the crash, but we still cannot see the web applications, but
a black area.

[SPEC-1934] WAM: crash in browser process GPU thread launching apps in qemux86_64