chromium68: use GCC for v8 toolchain if target is x86-64
This is a workaround for the problems trying to build for x86-64.
It was trying to build torque tool using target toolchain instead
of host toolchain.

Ideally we would want to use gcc for host in all cases, but it is
not possible in yocto as we do not have a way to obtain a native
toolchain for 32 bits targets.